migrated relivant tests to AttributeCard.spec.js
